requirements.txt
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic==0.25.0
python-dotenv==1.0.1